Add 'database create' commandline option.
This new feature was added in database section of commandline
options. This caused a change of the way the program behaves when
when the database is not available. Previously, the program
automatically created the database. After this change, the user has
to create the database using the new commandline feature or import the
database using the 'database import' commandline option.