remove duplicates when retrieving scheduled episodes.
since some shows can release multiple episodes at once, their episodes
can appear duplicated in the discover view as it is concerned with
displaying just the show posters and not individual episodes. This change
fixes that to some extent, show posters will still be duplicated if
they are not consecutive to each other as this depends on the
dedup_by_key method of a vector.