change the fetching method of seasons and episodes
previously, seasons were fetched separately and each season had to fetch
it's own episodes in a loop, which means the more episodes it has the
more it had request the api for those episodes. Since the api provides
one request method to retrieve all episodes of a series, this becomes
more efficient as whats left is to parse which episode belongs to what
season in that json file. Also this provides straight forward mechanism
for caching.