add identifiable series page message.
This Series Message is useful to make sure that the appropiate
series page receives it's exact series message. Since the series
page can be switched rapidly by the user, some of the commands
might be running in the background and my complete when a new instance
of series page has been opened updating it with wrong data.