Improve daily shows loading speed.
Since `series Troxide` already retrieves the daily full schedule and
caches it, there is no need of requesting daily global and local aired
series from the api and just load them from the full schedule cache
instead. And since the filtering of these shows will be done by the
program itself, daily shows will be more accurate as it wont rely
on the API's date which might be 1 day off in some situations.